% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-report.r
\name{meffil.normalization.report}
\alias{meffil.normalization.report}
\title{Generate report on normalization performance}
\usage{
meffil.normalization.report(
  normalization.summary,
  output.file = "normalization-report.md",
  author = "Analyst",
  study = "Illumina methylation data",
  ...
)
}
\arguments{
\item{normalization.summary}{Output from \code{meffil.normalization.summary}.}

\item{output.file}{Default = "meffil-normalization-report.html".
If the file extension is not .htm, .html, .HTM or .HTML then
output will be in markdown format.}

\item{author}{Default = "Analyst". Author name to be specified on report.}

\item{study}{Default = "Illumina methylation data". Study name to be specified on report.}

\item{...}{Arguments to be passed to \code{\link{knitr::knit}}}
}
\description{
Generate HTML file that summarises the normalization.
}
\examples{
\dontrun{

}
}
